Somewhat Broken
When clicking a website once, and a username once, and retrieving data
once, it works fine.
Other than that, a boat load of errors occur.
When the retrieve button is clicked, it seems only the website name is
passed to the method.
Username isn't being changed when it is changed after already being set
in the choice box.