Fatal Problems Fixed, Working as Intended
Problem was that the usernames ArrayList was not clearing when a new
website was picked in the choicebox.
Conditional added so that the usernames ArrayList will be cleared if it
is not empty.
Conditional added so that the requestedData ArrayList is cleared if it
is not empty. This is done because if a new item is picked in the
choicebox, it can be added to the appropriate index of the ArrayList.
Choiceboxes have been moved into nested code, this is a result of
exceptions being throw if the code was not nested.